Made the insert and remove functions operate return the number of collisions.
I thought that it would be easier to write tests if the output of the insert
and remove operations was the number of collisions detected while attempting
to perform that insert or remove operation. This allowed me to remove some bits
of the dlbf.h and the Makefile that I conidered to be kind of hacky.