Fixed a problem with the dlbf_insert function.
The dlbf_insert function was failing whenever two keys that were represented by
the same K bits were inserted into the Bloom filter. Initially, I wanted this
functionality because I wanted to be able to tell when duplicate keys were
inserted into the Bloom filter. I did not consider the case in which the same
K bits are required to be set to represent the same two keys or the case in
which all of the K bits required to represent a particular key are set due to
other keys in the Bloom filter. Now, the dlbf_insert function should return the
number of hash functions that will output a bit offset that falls within a non
collision-free region when applied to a particular key following the insertion
of that key.

Unfortunately, there is no way to detect that the duplicate insertion of a key
should not count as a collision, but I don't know if it's the Bloom filter's
responsiblity to keep track of that.