Fixed a typo in the dlbf_insert function.
Instead of incrementing the correct bit corresponding to the region in which
a collision was detected when a collision was detected, I was incrementing some
random bit in the array of cached bit offsets. I fixed the typo and added a
test to catch the same mistake again in the future.